Defer Lacros data wipe if in safe mode.
In safe mode, skip wiping of Lacros user data dir or clearing of migration
completion bit even if Lacros is disabled. This ensures that Lacros
being only temporarily does not result in an unintended user data loss.

check that Lacros user data dir is not wiped.

Test: Manually test by crashing Ash when a certain flag is enabled then
Bug: 1320589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3632280
Commit-Queue: Yuta Hijikata <ythjkt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1001248}

+++ chrome/browser/ash/crosapi/browser_loader.cc
@@ -123,12 +123,16 @@ bool CheckInstalledAndMaybeRemoveUserDirectory(
     return false;
 
   // Since we're already on a background thread, delete the user-data-dir
-  // associated with lacros.
+  // associated with lacros. Skip if Chrome is in safe mode to avoid deleting of
+  // user data when Lacros is disabled only temporarily.
   // TODO(hidehiko): This approach has timing issue. Specifically, if Chrome
   // shuts down during the directory remove, some partially-removed directory
   // may be kept, and if the user flips the flag in the next time, that
   // partially-removed directory could be used. Fix this.
-  base::DeletePathRecursively(browser_util::GetUserDataDir());
+  if (!base::CommandLine::ForCurrentProcess()->HasSwitch(
+          ash::switches::kSafeMode)) {
+    base::DeletePathRecursively(browser_util::GetUserDataDir());
+  }
   return true;
 }
